// packages/core/src/deployments.generated.ts
/** Generated by packages/core/scripts/sync-abis.mjs — do not edit by hand. */

import type { ChainAddresses } from "./types.js";

export const DEPLOYMENTS: Record<number, ChainAddresses> = {
  31337: {
    "chainId": 31337,
    "mockUSDC": "0x5fbdb2315678afecb367f032d93f642f64180aa3",
    "orgRegistry": "0xe7f1725e7734ce288f8367e1bb143e90bb3f0512",
    "treasury": "0x8a791620dd6260079bf849dc5567adc3f2fdc318",
    "escalationRouter": "0x2279b7a0a67db372996a5fab50d91eaa73d2ebe6",
    "governanceModule": "0x610178da211fef7d417bc0e6fed39f05609ad788",
    "spendCapPolicy": "0xdc64a140aa3e981100a9beca4e685f962f0cf6c9",
    "policyStack": "0xa513e6e4b8f2a923d98304ec87f64353c4d5c853",
    "whitelistPolicy": "0x9fe46736679d2d9a65f0992f2272de9f3c7fa6e0"
  },
  84532: {
    "chainId": 84532,
    "mockUSDC": "0x0000000000000000000000000000000000000000",
    "orgRegistry": "0x0000000000000000000000000000000000000001",
    "treasury": "0x0000000000000000000000000000000000000002",
    "escalationRouter": "0x0000000000000000000000000000000000000003",
    "governanceModule": "0x0000000000000000000000000000000000000004",
    "spendCapPolicy": "0x0000000000000000000000000000000000000005",
    "policyStack": "0x0000000000000000000000000000000000000006",
    "whitelistPolicy": "0x0000000000000000000000000000000000000007"
  },
};